+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.OpenGL/GLPrograms/LinesProgram.cs
@@ -2,6 +2,9 @@
 using OpenTK.Graphics;
 using OpenTK.Graphics.OpenGL;
 using System;
+#if NETCOREAPP || NET
+using OpenTK.Mathematics;
+#endif
 
 namespace ScottPlot.OpenGL.GLPrograms;