+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _Not yet on NuGet..._
 * SignalXY: Fixed bug introduced in the last version that caused off-screen data to throw an ascending value exception (#4261, #4286) @RFBomb @StendProg
 * Controls: Added strong naming by signing assemblies for the WPF, Maui, and Eto controls (#4295) @RFBomb
 * OpenGL: Improve behavior of plots when grid lines are rendered beneath plottables (#4298) @StendProg
+* OpenGL: Improve support for OpenGL controls on modern .NET using platform-specific OpenTK versions (#4301) @StendProg
 
 ## ScottPlot 5.0.39
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-08-02_